01-broken-access-control/rbac/requirements.txt
# web-security-snippets — 01-broken-access-control/rbac
# Python dependencies — pinned to exact versions
# Install: pip install -r requirements.txt

Flask==3.0.3
Flask-Login==0.6.3
PyJWT==2.8.0
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.2
fastapi==0.111.0
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
pydantic==2.7.4
